Better directory structure for packaging.
There are various (gitignored) files in the root directory of the project
that shouldn't be included in the packaged version of the extension: the
packaging key, ctags, .gitignore, etc.

The old Rakefile based packaging system no longer works, I now just need to
upload a zip of the relevant files to the Chrome Web store instead.