Fix Firefox Quantum compatibility.
Recent versions of Firefox, possibly but not definitely aligned with the
release of Quantum, haven't been working with the Choosy extension.

The root of the problem is that creating an iframe with an `x-choosy` URL in
the background script no longer triggers Choosy from Firefox.

The alternative was to open a new tab with the `x-choosy` URL, and
immediately close it again.

Unfortunately, the new Firefox code does not work in Google Chrome. The tab
opening callback fires too early, and even adding a delay (ranging from 0 to
100 ms) doesn't give Chrome time to activate Choosy. In the meantime, the
tab opening animation is quite jarring.

This commit therefore introduces branching logic for how we define the
function that opens Choosy. So much of the code is now separate branches,
that I question the usefulness of a single repository. It might be useful to
split them up again in future.